Add capnp wrapper for Wallet interface
Note: The CustomBuildField function in common-types is added in this
commit to serialize PKHash objects. It's not possible to write a generic
CustomReadField function for PKHash objects yet because PKHash doensn't
have span<byte> constructor, but if that is added, one could be written,
and the CustomReadField overload for PKHash could be dropped. (See
https://github.com/bitcoin/bitcoin/pull/10102#discussion_r784906151)