refactor: Add fs::PathToString, fs::PathFromString, u8string, u8path functions
There is no change in behavior. This just helps prepare for the
transition from boost::filesystem::path to std::filesystem::path by
adding explicit path <-> string conversions which are needed due to a
lack of a boost::filesystem::path::imbue equivalent in std::filesystem.
A detailed description of the problem and full solution can be found in
https://github.com/bitcoin/bitcoin/pull/20744#issuecomment-916627496.