indexes, refactor: Remove pruned data checking code
Large parts of this commit are move only. Suggest reviewing with --color-moved.

Add new interfaces::Chain::attachChain method and start moving index
sync logic out of indexing code to the node code in interfaces::Chain.

In this commit, just move prune-checking code that runs at the beginning
of sync (the code that checks for pruned blocks and triggers the "goes
beyond pruned data" error). In later commits, also move the
validationinterface handlers and sync thread out of the indexing code to
attachChain.

This commit does not change behavior in any way.