indexy # indexes, refactor: Move Rewind logic out of Rewind to blockDisconnected and ThreadSync
indexes, refactor: Move Rewind logic out of Rewind to blockDisconnected and ThreadSync

This main thing this commit code does is change rewind code to stop calling
`CustomRemove` in a loop, and instead call `blockDisconnected` in a loop, with
`blockDisconnected` calling `customRemove`. This makes the `blockDisconnected`
method functional, so the rewind loop can later be moved out of indexing code
entirely, and left to the node code.

This commit does not change behavior in any way.