indexy # indexes, refactor: Stop incorrectly calling Interupt() and Stop() in BaseIndex destructor
indexes, refactor: Stop incorrectly calling Interupt() and Stop() in BaseIndex destructor

This code was never correct and never used. Now that m_notification and
m_handler members have been added, the incorrect usage is easier to
detect and an assert can be added.

This commit does not change behavior in any way.