init: Allow -noproxy and -proxy="" setting values
This drops the "No proxy server specified. Use -proxy=<ip> or -proxy=<ip:port>"
error when "-noproxy" or "-proxy=" command line arguments are specified, and
just makes bitcoin connect and listen normally in these cases.

The error was originally added in https://github.com/bitcoin/bitcoin/pull/20003
to prevent a bare "-proxy" argument with no assignment from clearing proxy
settings. But it was implemented in an overbroad way breaking "-noproxy" and
"-proxy=" arguments as well.

The motivation for this change is actually to prevent a GUI bug that happens
with https://github.com/bitcoin/bitcoin/pull/15936, reported in
https://github.com/bitcoin/bitcoin/pull/15936#pullrequestreview-937685759 by
vasild, that happens after a proxy setting is enabled and disabled in the GUI.
But this change also makes sense on its own to remove a potentially confusing
error message, and make parsing of the "-proxy" setting handling consistent
within the InitParameterInteraction, AppInitParameterInteraction, and
AppInitMain functions.