refactor: Add util::Result failure values
Add util::Result support for returning more error information. For better error
handling, adds the ability to return a value on failure, not just a value on
success. This is a key missing feature that made the result class not useful
for functions like LoadChainstate() which produce different errors that need to
be handled differently.

This change also makes some more minor improvements:

- Smaller type size. util::Result<int> is 16 bytes, and util::Result<void> is 8
  bytes. Previously util::Result<int> was 72 bytes and util::Result<void> was
  unimplemented.

- Broader type compatibility. Supports Result<void> return values, so it is
  possible for functions that don't return values to still report errors.
  Supports noncopyable and nonmovable success and error types.