refactor: Replace BResult with util::Result
Rename BResult class to util::Result and update the class interface to be more
compatible with std::optional and with a full-featured result class implemented
in https://github.com/bitcoin/bitcoin/pull/25665. Motivation for this change is
to update existing BResult usages now so they don't have to change later when
more features are added to the Result class.