refactor, kernel: Remove gArgs accesses from dbwrapper and txdb
Code in the libbitcoin_kernel library should not be calling ArgsManager methods
or trying to read options from the command line. Instead it should just get
options values from simple structs and function arguments that are passed in
externally. This commit removes gArgs accesses from dbwrapper and txdb modules
by defining appropriate options structs, and is a followup to PR's #25290
#25487 #25527 which remove other ArgsManager calls from kernel modules.

This PR does not change behavior in any way.