Deduplicate bitcoind and bitcoin-qt init code
Add common InitConfig function to deduplicate bitcoind and bitcoin-qt code
parsing the config file, creating the datadir, and selecting the network.

There is a slight change of behavior for bitcoin-qt: When there is a problem
reading the configuration file, the error prefix is changed from "Error: Cannot
parse configuration file:" to "Error reading configuration file: ". The new
message is more accurate and makes the GUI consistent with the CLI.