kernel: Remove StartShutdown calls from validation code
This change drops the last kernel dependency on shutdown.cpp. It also adds new
hooks for libbitcoinkernel applications to be able to stop kernel operations
after blocks are imported and each time the chain tip changes.

This is a refactoring that mostly does not affect behavior, except it does
change the behavior of the -stopatheight option in a small way:

- Previously the -stopatheight option would shut down the node at or above the
  specified height. It would not stop connecting blocks after the specified
  height if they were already downloaded.

- Now, the -stopatheight option will shut down the node exactly at the
  specified height, without validating and connecting the downloaded blocks
  after it. Since blocks above the specified height are no longer validated, it
  means -stopatheight option is only guaranteed to return a valid chain with
  the specified height. But it is possible for the resulting chain to not be the
  most-work chain, if the blocks above the height which are not validated turn
  out to actually be invalid.